Add logging formatters (#18)
- We need structured logs for later error/log analysis --> JSON a widely
supported standard.
- Logged to file & terminal via custom formatters
- JSON & terminal & console logs have the same milli second timestamp to
make logs searchable/joinable
The Rich Handler does not support writing to files (I did not find any
option). Manually redirecting to file is a bit ugly because it uses a
lot of special terminal characters for the colors and formatting.
Therefore, I created my own custom formatter with colors. We use
better-exceptions to nicely add more information to error traces. It
supports turning off color for file based logging.

- Removed rice progress bar because it is not nicely compatible with
logging. We can reintroduce that once we move to a nice interface for
"customers"
- Modified ontology helper to not log multiple times in init